refactor(tests): move utils unit tests to their own module
