+++ tests/unit/utils/test_file.py
@@ -1,0 +1,145 @@
+# Copyright 2023 Canonical Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# For further info, check https://github.com/canonical/charmcraft
+"""Unit tests for file-related utilities."""
+import os
+import pathlib
+import sys
+import zipfile
+
+import pytest
+from craft_cli import CraftError
+
+from charmcraft.utils.file import build_zip, make_executable, useful_filepath
+
+
+@pytest.mark.skipif(sys.platform == "win32", reason="Windows not [yet] supported")
+def test_make_executable_read_bits(tmp_path):
+    pth = tmp_path / "test"
+    pth.touch(mode=0o640)
+    # validity check
+    assert pth.stat().st_mode & 0o777 == 0o640
+    with pth.open() as fd:
+        make_executable(fd)
+        # only read bits got made executable
+        assert pth.stat().st_mode & 0o777 == 0o750
+
+
+def test_usefulfilepath_pathlib(tmp_path):
+    """Convert the string to Path."""
+    test_file = tmp_path / "testfile.bin"
+    test_file.touch()
+    path = useful_filepath(str(test_file))
+    assert path == test_file
+    assert isinstance(path, pathlib.Path)
+
+
+@pytest.mark.skipif(sys.platform == "win32", reason="Windows not [yet] supported")
+def test_usefulfilepath_home_expanded(tmp_path, monkeypatch):
+    """Home-expand the indicated path."""
+    fake_home = tmp_path / "homedir"
+    fake_home.mkdir()
+    test_file = fake_home / "testfile.bin"
+    test_file.touch()
+
+    monkeypatch.setitem(os.environ, "HOME", str(fake_home))
+    path = useful_filepath("~/testfile.bin")
+    assert path == test_file
+
+
+def test_usefulfilepath_missing():
+    """The indicated path is not there."""
+    with pytest.raises(CraftError) as cm:
+        useful_filepath("not_really_there.txt")
+    assert str(cm.value) == "Cannot access 'not_really_there.txt'."
+
+
+@pytest.mark.skipif(sys.platform == "win32", reason="Windows not [yet] supported")
+def test_usefulfilepath_inaccessible(tmp_path):
+    """The indicated path is not readable."""
+    test_file = tmp_path / "testfile.bin"
+    test_file.touch(mode=0o000)
+    with pytest.raises(CraftError) as cm:
+        useful_filepath(str(test_file))
+    assert str(cm.value) == f"Cannot access {str(test_file)!r}."
+
+
+def test_usefulfilepath_not_a_file(tmp_path):
+    """The indicated path is not a file."""
+    with pytest.raises(CraftError) as cm:
+        useful_filepath(str(tmp_path))
+    assert str(cm.value) == f"{str(tmp_path)!r} is not a file."
+
+
+def test_zipbuild_simple(tmp_path):
+    """Build a bunch of files in the zip."""
+    build_dir = tmp_path / "somedir"
+    build_dir.mkdir()
+
+    testfile1 = build_dir / "foo.txt"
+    testfile1.write_bytes(b"123\x00456")
+    subdir = build_dir / "bar"
+    subdir.mkdir()
+    testfile2 = subdir / "baz.txt"
+    testfile2.write_bytes(b"mo\xc3\xb1o")
+
+    zip_filepath = tmp_path / "testresult.zip"
+    build_zip(zip_filepath, build_dir)
+
+    zf = zipfile.ZipFile(zip_filepath)
+    assert sorted(x.filename for x in zf.infolist()) == ["bar/baz.txt", "foo.txt"]
+    assert zf.read("foo.txt") == b"123\x00456"
+    assert zf.read("bar/baz.txt") == b"mo\xc3\xb1o"
+
+
+@pytest.mark.skipif(sys.platform == "win32", reason="Windows not [yet] supported")
+def test_zipbuild_symlink_simple(tmp_path):
+    """Symlinks are supported."""
+    build_dir = tmp_path / "somedir"
+    build_dir.mkdir()
+
+    testfile1 = build_dir / "real.txt"
+    testfile1.write_bytes(b"123\x00456")
+    testfile2 = build_dir / "link.txt"
+    testfile2.symlink_to(testfile1)
+
+    zip_filepath = tmp_path / "testresult.zip"
+    build_zip(zip_filepath, build_dir)
+
+    zf = zipfile.ZipFile(zip_filepath)
+    assert sorted(x.filename for x in zf.infolist()) == ["link.txt", "real.txt"]
+    assert zf.read("real.txt") == b"123\x00456"
+    assert zf.read("link.txt") == b"123\x00456"
+
+
+@pytest.mark.skipif(sys.platform == "win32", reason="Windows not [yet] supported")
+def test_zipbuild_symlink_outside(tmp_path):
+    """No matter where the symlink points to."""
+    # outside the build dir
+    testfile1 = tmp_path / "real.txt"
+    testfile1.write_bytes(b"123\x00456")
+
+    # inside the build dir
+    build_dir = tmp_path / "somedir"
+    build_dir.mkdir()
+    testfile2 = build_dir / "link.txt"
+    testfile2.symlink_to(testfile1)
+
+    zip_filepath = tmp_path / "testresult.zip"
+    build_zip(zip_filepath, build_dir)
+
+    zf = zipfile.ZipFile(zip_filepath)
+    assert sorted(x.filename for x in zf.infolist()) == ["link.txt"]
+    assert zf.read("link.txt") == b"123\x00456"
